Update autoMod.js
Doing some more trouble shooting steps

+++ events/messageCreate/autoMod.js
@@ -13,19 +13,21 @@ module.exports = async (client, message) => {
 
     if(!data || !data.autoModEnabled) return;
 
-    const youtubeRegex = /(?:https?:\/\/)?(?:www\.)?(?:youtube\.com|youtu\.be)\/(?:watch\?v=)?(.+)/gi;
-    const discordRegex = /(discord\.gg\/|discordapp\.com\/invite\/)/gi;
-    const twitchRegex = /(?:https?:\/\/)?(?:www\.)?(?:twitch\.tv)\/(.+)/gi;
-    const racialSlursRegex = /(?:n[\*\^]{1,}g{1,}a{0,1}|b[\*\^]{1,}tch|f[\*\^]{1,}g{1,}|s[\*\^]{1,}t)/gi;
-    const messageContent = message.content.toLowerCase();
-
-    if(youtubeRegex.test(messageContent) || discordRegex.test(messageContent) || twitchRegex.test(messageContent)) {
-        await message.delete();
-
-       await message.channel.send(`${message.author}, you can't sned links here!`);
-    } else if(racialSlursRegex.test(messageContent)) {
-        await message.delete();
-
-       await message.channel.send(`${message.author}, you can't say that word here.`);
+    if(data || data.autoModEnabled) {
+        const youtubeRegex = /(?:https?:\/\/)?(?:www\.)?(?:youtube\.com|youtu\.be)\/(?:watch\?v=)?(.+)/gi;
+        const discordRegex = /(discord\.gg\/|discordapp\.com\/invite\/)/gi;
+        const twitchRegex = /(?:https?:\/\/)?(?:www\.)?(?:twitch\.tv)\/(.+)/gi;
+        const racialSlursRegex = /(?:n[\*\^]{1,}g{1,}a{0,1}|b[\*\^]{1,}tch|f[\*\^]{1,}g{1,}|s[\*\^]{1,}t)/gi;
+        const messageContent = message.content.toLowerCase();
+    
+        if(youtubeRegex.test(messageContent) || discordRegex.test(messageContent) || twitchRegex.test(messageContent)) {
+            await message.delete();
+    
+           await message.channel.send(`${message.author}, you can't sned links here!`);
+        } else if(racialSlursRegex.test(messageContent)) {
+            await message.delete();
+    
+           await message.channel.send(`${message.author}, you can't say that word here.`);
+        }
     }
 };
